Move weather queries to their own class.
Pull out the code that does the weather API queries to a new
modeule with a new class.

The idea is that this will allow the main module to be more
of a template that can be used for all weather node servers and
have the weather service specific parts isolated in a single
class.